+++ src/Exceptions/ShouldBeSkippedMetricExceptionInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace AvtoDev\AppMetrics\Exceptions;
+
+interface ShouldBeSkippedMetricExceptionInterface extends \Throwable
+{
+    //
+}